Addressed PR comments with the TabBars, renamed file
