+++ src/components/common.js
@@ -303,7 +303,7 @@ const formatGroupLabel = group => (
     }
   }, [group.label]))
 
-const BaseSelect = ({ value, newOptions, id = null, findValue, maxHeight, ...props }) => {
+const BaseSelect = ({ value, newOptions, id, findValue, maxHeight, ...props }) => {
   const newValue = props.isMulti ? _.map(findValue, value) : findValue(value)
   const [isOpen, setOpen] = useState(false)
   const menuId = Utils.useUniqueId()
