+++ src/components/data/data-utils.js
@@ -9,7 +9,7 @@ import Dropzone from 'src/components/Dropzone'
 import { icon } from 'src/components/icons'
 import { NumberInput, PasteOnlyInput, TextInput, ValidatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
-import { SimpleTabBar } from 'src/components/TabBar'
+import { SimpleTabBar } from 'src/components/tabBars'
 import { TextCell } from 'src/components/table'
 import TooltipTrigger from 'src/components/TooltipTrigger'
 import { UriViewerLink } from 'src/components/UriViewer'
