+++ src/pages/workflows/List.js
@@ -6,7 +6,7 @@ import { AutoSizer } from 'react-virtualized'
 import { Link } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { DelayedSearchInput } from 'src/components/input'
-import { TabBar } from 'src/components/TabBar'
+import { TabBar } from 'src/components/tabBars'
 import { FlexTable, HeaderCell, Sortable, TooltipCell } from 'src/components/table'
 import TopBar from 'src/components/TopBar'
 import { Ajax } from 'src/libs/ajax'
