+++ src/pages/workspaces/List.js
@@ -10,7 +10,7 @@ import { icon } from 'src/components/icons'
 import { DelayedSearchInput } from 'src/components/input'
 import NewWorkspaceModal from 'src/components/NewWorkspaceModal'
 import PopupTrigger from 'src/components/PopupTrigger'
-import { SimpleTabBar } from 'src/components/TabBar'
+import { SimpleTabBar } from 'src/components/tabBars'
 import { FlexTable, MiniSortable } from 'src/components/table'
 import TooltipTrigger from 'src/components/TooltipTrigger'
 import TopBar from 'src/components/TopBar'
