+++ src/pages/workspaces/workspace/WorkspaceContainer.js
@@ -8,7 +8,7 @@ import { icon } from 'src/components/icons'
 import NewWorkspaceModal from 'src/components/NewWorkspaceModal'
 import PopupTrigger from 'src/components/PopupTrigger'
 import RuntimeManager from 'src/components/RuntimeManager'
-import { TabBar } from 'src/components/TabBar'
+import { TabBar } from 'src/components/tabBars'
 import TopBar from 'src/components/TopBar'
 import { Ajax, saToken } from 'src/libs/ajax'
 import { getUser } from 'src/libs/auth'
